nmci.process: don't accept default values for arguments in nmci.process._run()
The underscore indicates that this is an internal function. If you
call it (which you probably should not), you should also provide all
the arguments.

Note that this function has wrappers, which do accept default values
(which differ, as it makes sense for them). Having two layers of default
values is confusing.