update python black version
Newer versions format code slightly different. In particular, they
nicely interpret a trailing comman in an argument list, to wrap
the lines and put each argument on a separate line.

Also, on our developer machines, we wouldn't have such an ancient
black version. It's kinda important that everybody uses the same
black version.

In practice, it was so far always possible to modify the code so that
older and recent black versions format it the same. But by updating
to a newer version, we no longer need that.