adjust version tags for test variants for checking order of IPv6 addresses
With [1] we change the order for which we configure IPv6 addresses.
This is an intentional change in behavior compared to older versions.

The change already hit `main` branch first after 1.39.2 (@ver+=1.39.2).
It was also already backported to nm-1-38 (after 1.37.91) to make it into
upstream 1.38.0 release (@ver+=1.37.91).

The change will soon also make it into Fedora 36, which is still getting
rebased to 1.38.0 (and 1.38 release candidates). The behavior for
Fedora will correspond to the upstream versioning.

The change for the moment will be reverted on RHEL (this affects now rhel-8.7
and rhel-9.1, which will be rebased to upstream 1.39+, but without this
change.

Whether we change behavior in future RHEL (rhel-8.8+ and rhel-9.2+) is
not yet decided. RHEL-10 will get the new behavior, unless there are
major problems so we fix upstream behavior again.

Adjust the version tags.

[1] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/c3b7ec5b4258b9dd932fd85d3dc0f90400097146