nmci: cleanup @rhel_pkg in misc.test_tags_match_version() and add code comment
nm_stream.startswith("rhel") is not great, because the stream is
hierarchical, where "-" separates the parts.

In practice, the stream gets set by misc.nm_version_parse(), so the
previous code before was correct for all inputs we would get. Still, we
should match here against the first element of the stream (that is, the parts
before "-", if any).

Also add a code comment about the relation of the @rhel_pkg,
@not_with_rhel_pkg, @fedora_pkg, @not_with_fedora_pkg tags and
the versioned @ver/rhel, @ver/fedora tags. There is an overlap here.
For example,

  "@ver+=1.35 @ver-1.38 @rhel_pkg" would be like "@ver- @ver/rhel+=1.35 @ver/rhel-1.38"

But there is also a difference. @rhel_pkg et.al also take into account
the distro-flavor, while @ver/rhel only cares about the package
"stream", but not where you run. That could make a difference if
you recompile a rhel package and run on Fedora.