nmci/tests: extend test for validating version tags
Ensure that we can completely re-construct the full tag name based on
the information that we parsed.

The version tags that we use should be written in a unique way.
That is, we want "@ver+=1.3" and reject "@ver+=1.03".

The purpose is that we might want to grep for them, and then there
should be only one way to express a certain tag.