fix @ipv6_check_addr_order test for 1.36
This test was added to find out what the current behavior is. That is, I
didn't start with what I think should happen, instead I tried to write
the test so that it passes the current code. The idea was to find out
the currently implemented behavior, and to codify/fix it with a test.

Apparently this test is failing on nm-1-36, so the logical solution is to
adjust the test so that it passes and not to think why it behaves
differently than I thought it would.

This means, that the test for [1.35..1.37.89] is now the same as for
[1.37.90..1.37.91] and they can be merged.