ipv6: fix @ipv6_check_addr_order for rhel-8.6/rhel-9.0 packages
We run the test against upstream nm-1-36 branch, let's also enable it
(again) for the respective RHEL packages. But note that the nm-1-36 branch
behaves different from the RHEL packages. Possibly this was introduced some
time after 1.36.0 on a minor release, while RHEL didn't get that fix.
The exact version is not really known, but we won't test upstream
releases like 1.36.2 anymore (only latest nm-1-36 branch). So the
difference doesn't matter.

The goal here is not to question why RHEL packages behave different, but
to have a test that codifies the actual behavior.