nmci: add mechanism for temporary dir in NM-ci tree
We should not put random stuff in /tmp. When we run tests, we
sometimes leave state on disk. But it should be possible to wipe
all changes by deleting one directory. If any number of files are
created in /tmp, we don't know which files are relevant.

Introduce nmci.util.tmp_dir() which gives (and creates) a temporary
directory at a well known location: "./.tmp/".