nmci: support overwriting stdout/stderr with nmci.process.run()
nmci.process.run() was supposed to be a simple wrapper around
subprocess.run(). Well, turns out, it is actually pretty useful to be
able to redirect the output. So add support for that too.

Note there is an argument, to keep nmci.process.run()'s API simpler, and
users that need to redirect output, should instead use subprocess.run()
directly. However, nmci.process.run() is hooked up to the context as
context.process.run(), which does additional things like logging the
call. What nmci.process.run() still provides over subprocess.run(),
is that it defaults to fail on any unexpected things (e.g. the process
being killed by a signal).

So add it. But only to nmci.process.run(), and not the simpler wrappers
like nmci.process.run_code(). Use wisely and sparsely.