nmci: add "nmci.util.ExpectedException"
Usually, we want to fail-fast. That means, we should not try to catch
any unexpected exceptions. This is not driving a nuclear reactor which
must try to recover from unexpected failure. Instead, every unexpected
failure needs to be propagated and not caught.

Sometimes however, we use exception to signal expected failures. In
those cases, `raise Exception()` is bad, because the caller is then
forced to catch all exceptions, even the unexpected ones. Usually
the right solution is to raise an context specific exception type.

However, then we might end up adding a lot of exception classes. That
is also not very useful. Instead, add a special "nmci.util.ExpectedException"
type, which fills this purpose. We should raise this exception, when we
expect the caller might want to catch it. And callers should avoid
`except Exception`.