avoid race in "@modify_policy_based_routing_connection"
Test "@modify_policy_based_routing_connection" activates a profile with
IPv4 and IPv6 auto methods enabled (and `ipv4.may-fail=yes`). This
means, the device might be already considered up without an IPv4
address.

Afterwards, the test calls

  * Create PBR files for profile " con_general " and " testG " device in table " 1 "

which then tries to get the IPv4 address. If it fails, because the
address is not yet configured the test will fail.

Let the "Create PBR files for profile ..." step by default retry for up
to 5 seconds to get a valid IPv4 address.

https://bugzilla.redhat.com/show_bug.cgi?id=2079409#c1