nmci/lib: rename module to "nmci.ctx"
This module is about the behave context. Give a better name.
"lib" means anything. In particular, we already have "nmci.util"
and "nmci.misc" which too are unspecified heaps of utils. "nmci.lib"
sounds as it would be of that kind. But it isn't and it actually
has a clear purpose: the behave context. Rename.