nmci/ctx: add "context._ctx_ext" field
nmci.ctx.context_setup() and various other functions add additional
attributes to the behave context object. While python is a dynamic
language and you can technically do this, it means it becomes confusing
which part of the context are original behave parts, and which are our
own additions.

Add a _CtxExt class, and let nmci.ctx.context_setup() set that as
"context._ctx_ext". All our extensions should be on our own object,
and not pollute the naming of the behave context.

This is really just to add a namespace for our own stuff.

Additionally, now that we have an object, we can have actual functions
defined in the _CtxExt class. If we extend the behave context, we could
only simulate functions by setting object members to functions. Of
course, python has also functional language paradigms, so setting
a field to a function is not inherently wrong. Still, we will see that
is just looks nicer to have proper object functions.