tags: wifi fixes
wifi_as was disabled, but it is needed since wifi feature sometimes use 
`nmcli device wifi connect`. Enabling wifi_as again. Better solution 
would be to use cleanup step in wifi scenarios, but that feature is 
planned to be completely rewritten to the new lab, so this solution is 
best effort now.