nmci/ctx: add and use context.cext.embed_data() wrapper
embed() basically overwrites the embed function of
behave-html-formatter. Which accepts a mime_type argument.
NM-ci would also hack the mime-type to support "call" and
"link" types. Those are specially interpreted by NM-ci (although, "link"
is also passed on to the html-formatter's embed() function).

Having one embed function that accepts different arguments that depend
on the "mime_type" parameter is hard to understand. What are even the
valid mime_types?

Add an embed_data() function which straight forward just passes the
parameters to the behave formatter.

Later, we will use this to cleanup further.