nmci/ctx: print embedded data to stdout if requested
when run in the terminal, all embeds were ignored. I think it can be
useful to see them for debugging.

Unfortunately, for regular command line use is this output too
verbose/confusing, so only show it with NMCI_SHOW_EMBED=1.

Also, a run with or without html formatter should behave similar.
Suppressing all embeds early on results in different behavior, which
could lead to bugs. E.g. when you expect that the callback of
embed_later() always gets invoked.