nmci/ctx: always attach data at the end during embed()
Many embed items are "fail_only=True", "call" type, or have a
combine_tag. Those can only be attached at the end of the run,
when we have all the information.

Some items could be attached right away, however, if we do that
it makes the code slightly more complicated. But worse, it means
to order in which we attach embeds can be destroyed.

Instead, always attach all embeds at the end.