nmci/ctx: always embed parts of the attachment even on success
Attaching all the data can result in large HTML artifacts (megabytes).
Hence, we tend to only attach the data, if the test failed.
For that there is the "fail_only" property.

However, the embeds also give a better picture what is happening,
so omitting on successful runs is painful.

Instead, always attach at least the caption and 2k of the message (with
a note that it was truncated).