nmci/ctx: run process during dump_status() in parallel
We call dump_status() up to 3 times per test and it takes a few
hundreds of milliseconds. As we run 1000+ tests, this easily
amounts to minutes.

Run the commands in parallel.

While at it, also make use of "shell=nmci.process.SHELL_AUTO", that is,
to encode whether to use a shell in the first character of the command.
Several of the commands didn't need to go through the shell.