editor: add steps 'Check regex "{regex}" in describe output for object "{obj}"'
We already have two steps like
  'Check "{options}" are present in describe output for object "{obj}"'
  'Check "{options}" are shown for object "{obj}"'

However, the "options" are a list of regexes, separated by "|". That means,
"|" cannot be used in a regex anymore, which is a severe limitation.

The former has legitimate uses, like:

  Check "mac-address|stp|priority|forward-delay|hello-time|max-age|ageing-time" are present in describe output for object "bridge"

However, there are also legitimate uses of using a regex with "|", like:

  Check regex "=== \[dns-search\] ===\s+\[NM property description\]\s+(List|Array) of DNS search domains." in describe output for object "dns-search"

Preserve the original variants in their existing form, because it's not
clear to me how to preserve the functionality while consolidating it
with true regexes. Instead, add two new variants that take plain
regexes.