prepare: don't bring up simulated devices
Bringing up our end of the Veth pair is not a good idea, with dnsmasq
sending RAs and the kernel accepting them, the devices might end up
being externally configured. This typically not what we'd want during
testing.

Specifically, this interferes with the @ovs_cloned_mac_set_on_iface.
The test creates a testX device and an ovs-testX connection for the
device. It expects it to autoactivate, but that doesn't always happen
because the device might end up being externally configured after
getting a RA from dnsmasq.

That results in a truly sad state of affairs:

  --- nmcli -f ALL d ---
  DEVICE  TYPE      STATE                   CONNECTION
  testX   ethernet  connected (externally)  testX

  --- nmcli -f ALL c ---
  NAME       TYPE      ACTIVE  DEVICE  STATE      SLAVE
  testX      ethernet  yes     testX   activated  --
  ovs-testX  ethernet  no      --      --         ovs-port