gitlab-ci: run "UnitTests" test in a separate stage "unit"
Previously, the tests "UnitTests" and "TestResults" are both in the same
stage "test". Note that we had in total two stages: "test" and "external".
And in the web UI, even before clicking on the pipeline, you would see the
overall result of the stage (red or green).

Now, the unit tests usually would pass, but the "external" tests (and thus
"TestResults" might fail. As a result, the entire stage was marked as failed.
That is annoying. I care whether the unit tests pass, and that should be visible
without clicking on the pipeline.

Move it to a separate stage.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager-ci/-/merge_requests/1096