nmci/util: add "loop_sleep()" method to nmci.util.start_timeout() helper
This allows for a convenient pattern:

    timeout = nmci.utils.start_timeout(10)
    while timeout.loop_sleep(0.1):
        if ready():
            break
    if timeout.was_expired:
        raise Exception()

loop_sleep() ensures that the loop is run at least once. Even with
a timeout of "0". It also ensures that the last sleep before timeout
is shortened to not go beyond the expiry.

Also, accept "timeout" as string. This is convenient when we get the
timeout from a behave step argument.