device: make sure new device gets the highest ifindex
Some tests rely on the newly appearing device getting the highest
interface index.  This is not always the case. In particular, it might
not be the case if we create a device in a separate namespace (which a
follow-up commit is going to do) and then we move it.

Let's be explicit about the ifindex we expect.

This will also catch a failure in an unlikely event of ifindex
wraparound.