steps: add loading and clean-up of nftables
step also prints ruleset in respective namespace before and after
loading the new ruleset.

nftables rulesets are also printed in dump_status() now (producing
one line per NS when nftables are not touched) and cleanup that flushes
nftables in respective namespace if it still exists and it is touched by
this step or if requested by new Cleanup nftables step.

fixes: https://gitlab.freedesktop.org/NetworkManager/NetworkManager-ci/-/issues/49