prepare: drop "envsetup" directory and add all the code to "prepare/envsetup.sh"
Code does not automatically get easier to understand by splitting it into
a larger number of files. In particular, this code was sourced by
"envsetup.sh". These snippets where thus not self-contained and could
not be meaningfully used alone or for another purpose. Despite being
in multiple files, the "envsetup.sh" code is effectively one very large
shell script spread out over multiple files.

We don't need this. We also don't have a plugin architecture, where externally
maintained plugins could extend the workings of "envsetup.sh" by being sourced
(nor would that be desirable, as it increases the complexity tremendously and
there simply is no need).

Well, maybe "pkg_install_el7.sh" has a clear purpose, but since
RHEL/Fedora versions are similar, it means we had several entirely
separate installation routines. That is bad, because all RHEL/Fedora-like
systems are similar, so it would be much better to have one installation
routine that works for all of them, and have small conditionals for
the places where there are differences. Instead, the installation
routines were effectively forked, similar yet diverging. It was hard
to understand whether something important is done for all the versions
where it's necessary, and where the differences are.

Drop the files and move all the code to "envsetup.sh". With this we can
see in one file what is happening. Otherwise, you see just a call to
install_packages(), which is defined somewhere else (which expands to 37
lines of shell) and then calls "install_"$release"_packages", which is
defined somewhere in 4 different files.

Having it all in one place makes it simpler to reason about it, and
most importantly, it opens opportunities for unifying things and
drop redundant code (which is there).

This is also, because we want to have even more environments, like
running in a container. Things should be unified before adding more
code.