envsetup: add and use "nmci/sh/nmci-utils.sh" script
We often source other shell scripts. That is usually bad, because
it's hard to understand what happens then (without fully understanding the
sourced script). It's much better to execute scripts, because then they
run in a separate context and their API (command line arguments and/or
environment variables) are simpler and not an entire programming language.

Also, we both execute and source "prepare/envsetup.sh" script, which is
generally a bad idea. Either source it or execute it. Doing it both is
odd.

However, there is a point in having *one* file that can be sourced. So add one:
"nmci/sh/nmci-utils.sh". This file is supposed to provide helpers for our
other NM-ci shell scripts can can be freely sourced by them. It also extends
the $PATH and sets $NMCI_BASE_DIR.

It's unlikely that there will be another such utils script that is to be
sourced. Aside from "nmci/sh/nmci-utils.sh", we should aim to add
self-contained scripts that get executed. Those would then reside under
"nmci/helpers" directory.