prepare: rework overwriting python command
First, we should stop sourcing "prepare/envsetup.sh".
One reason why it was sourced was to get "export_python_command",
so this needs to move somewhere else.

That new place is "nmci/sh/nmci-utils.sh".

While at it, change the mechanism how this works. Instead of
exporting a bash function, drop a wrapper script to ".tmp/bin"
(which is in the $PATH).

I wonder whether this is actually working. We only do this for rhel7,
which is no longer tested. Also, we execute several python programs
directly, where they run the python from the shebang and don't honor
the override.