prepare: don't source envsetup.sh and simplify command line arguments
A script that can both be sourced and executed is odd. Don't do that.
It was only sourced in one place (runtest.sh), we can just execute it.

Drop the "setup" argument. Only that one was supported anyway.

Also, don't use the special word "setup first_test_setup" to indicate
to perform only the test-independent setup. Instead, don't specify
the test name, if we don't want the test-dependent setup.