envsetup: drop f32 workaround for sshd bug
Even in f32 is this fixed, and a plain `dnf update` will give a newer
sshd than the one that we tried to install. This code basically
will never install a package, because the installed package is newer
already. Not to mention, that this ever was only relevant on Fedora 32.