nmci: cleanup pexpect services via the cleanup mechanism
We now have a cleanup mechanism, which most importantly, supports
specifying the order in which cleanup happens.

Cleanup the processes from context.pexpect_service() via that mechanism.
This allows to schedule a cleanup action that runs before/after reaping
the process, and you know it because you are in control of the order.

context.pexpect_spawn() are still reaped the same was as before.

As we already track the spawned processes, add a label property so we
can find them again. This is used by the "Run child" steps. Those just
use `label=True`, but the label can be anything that can be compared.

While at it, use name CExt.pexpect_{spawn,service}(). You can now
officially call the code via `context.pexpect_*()` or via
`context.cext.pexpect_*()`. It's the same thing, whereas the former
may be more convenient to write, but is really just an alias.