nmci: explicitly write arguments to pexpect_{spawn,service}
When reviewing the code, we want to see what arguments exist and
how we use them.

Previously, we had some arguments that were interpreted by us (`shell`),
some that were specified by as and passed on to pexpect.spawn()
(`encoding`) and any other arguments (`*a, **kw`) which were blindly
passed to pexpect.spawn(). This makes it hard to review the code
and understand what exact arguments are supported.

Check the parameters we currently use, and list those explicitly.
Yes, this does not list all parameters that pexpect.spawn() supports.
If you need more, add them.

I am fine with passing on (*a, **kw), if we would pass them on entirely.
But mixing our own parameters and parameters that are handed to pexpect
is hard to reason.