nmci: drop unused context.cext.embed_later()
It was not actually used. But also, you can now achieve the same with
cleanup actions.

Granted, embed_later() ran *very* late, so to implement that with
cleanup actions, we probably need to carefully craft the things
that run between process_cleanup() and process_embeds(), so that
it's late enough.

Anyway, it was unused.