envsetup: explicitly call distro-dependent install functions
These functions will be merged in the next step. Call them explicitly in
preparation for that.

Also drop the retry. It had only a naive check for whether some packages
were installed, but did not cover everything. That means, it may only
retry in some cases, not in many other. If really necessary, what we
would need is that the individual steps retry accordingly and apply
the necessary workarounds (e.g. nmci_install). However, that does not
necessarily seem right. If there is temporarily a networking glitch,
then that needs fixing, and if it still happens, the test should just
fail. We also have tests that check connectivity, so on such a system
those tests would randomly fail, which is a no go.