environment: call process_pexpect_spawn() earlier
Stop per-step processes early during cleanup, before running the
steps' after-scneario callbacks. Since per-process steps are
also reaped at the end of the step, it's not clear this is
needed at all. In any case, it probably should happen earlier.

Also, drop the try/except. Other calls of this function don't
have it either, and this function is anyway not supposed to raise
an exception.