ctx: schedule cleanup action to start NetworkManager service
Test @shutdown_service_connected creates a profile "con_general"
and then stops NetworkManager service.

The cleanup action fails to delete the profile, because the service is
not running. Instead, ensure that when stopping NetworkManager we
restart it early during cleanup.

Clearly we always want that NetworkManager gets restart at the end of
the scenario. For that reason, we could also just hard code that in the
after scenario. Instead add a cleanup action because it has priorities
and we can schedule other actions before/after.

https://jenkins-networkmanager.apps.ocp.ci.centos.org/job/NetworkManager-test-mr/2581/artifact/report_NetworkManager-ci-M0_Test0425_shutdown_service_connected.html