cico: use duffy for system reservation and returning
We directly reserve a session and number of machines we want. This
should avoid starving when one machine out of two is reserved and
the we are waiting for second blocking the first one. Build now
doesn't use another machine (as machine number limit is 10). but is
performed on the machine[0].
Returning machine is left to groovy script and returns whole session
with all machines at once.

Yes, we have public hostname again.