ipv6: add test for IPv6 DAD on stable-privacy SLAAC addresses
Add a test to check that when DAD fails for the stable-privacy IPv6
address generated from SLAAC, a new address is tried.

Since the stable-privacy interface identifier depends on the UUID and
the machine-id, it's not easy to predict what the IPv6 address will
be. So, to generate a collision, we first activate the connection and
look at what address is generated. Then we add the address in the
namespace and reactivate the connection.

See: https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/1366