nmci.misc: extend detection of NetworkManager version for test packages
When we create a test build for a RHEL package, it makes sense to add some
suffix to the version number, like "1.36.0-9.rh1000000.1.el8_6".

Fix the parsing to accept such suffixes too. While these are not the "proper"
RHEL packages that we actually ship, they are based on RHEL packages and their
version stream should honor that.