bridge_detect_initrd_device: use conn name different from dev name
Make is explicit the bridge ports refer the bridge by connection name,
not the device name.

If both are the same, nmcli doesn't resolve the name to the UUID.
That's inconsistent with what happens in initrd, but also ambiguous
enough for NM not to be able to match the assumed connection when [1] is
applied.

[1] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/1408