vethsetup: ensure processes are gone during cleanup in setup_veth_env
NetworkManager.service contains `KillMode=process`. With `[main].dns=dnsmasq`,
that means we keep dnsmasq running during `systemctl stop NetworkManager`.

  NetworkManager.service: Found left-over process 953666 (dnsmasq) in control group while starting unit. Ignoring.

That is somewhat intended, so that you can restart NetworkManager and
dnsmasq keeps running. However, after NetworkManager restarts, it will
see the PID file in `/run/NetworkManager/dnsmasq.pid`, kill dnsmasq and
restart it.

Well, at least that's how it is supposed to be. "vethsetup.sh" stops
NetworkManager, keeps dnsmasq running, and then prunes
/run/NetworkManager (and the pid file). NetworkManager then won't know
to stop dnsmasq, and a new dnsmasq instance cannot start because the PID
is in use.

NM-ci really should not hackishly call `systemctl stop` and `rm -rf
/run/NetworkManager`. Instead, there should be one implementation (a
script?) that does all the necessary steps for simulating a reboot of
NetworkManager. Necessary steps then could include to kill the entire
process group, or to kill the processes based on the PID files, like

  pkill -F /run/NetworkManager/dnsmasq.pid
  rm -rf /run/NetworkManager

But it's not done this way. Instead, just hack a `systemctl kill`
here. All other places that may do something similar are not fixed.

This fixes and issue that breaks follow-up tests with the loopback feature branch.
On that branch, a new device appears active ("lo"), and the cleanup happens at the
end of test @dns_dnsmasq_driver_removal. The following test @dns_dnsmasq_kill then fails
because dnsmasq cannot start.