dispatcher: delete dispatcher script via cleanup actions
Move logic away from @disp tag. We should use tags less, they are very
"detached" from the code.

E.g. the @disp tag only cleans up dispatcher scripts with name "*-disp".
Test @dispatcher_interface_stuck_in_check_ip_state does:

  * Write dispatcher "00-dhcp-dns.sh" file with params

so the dispatcher script was never cleaned up, and keeps running for
the rest of the test-run (producing warnings in the NM-log).

The problem is that "@disp" tag has some non-obvious interdependencies
with the "Write dispatcher" step. This is wrong. All the necessary code
should be in the behave step, including the (scheduling of the) cleanup.