ipv4: adjust version tags in @ipv4_routing_rules_manipulation test
The change was done in 1.41.4+ ([1]) and backported to 1.40.3+ ([2]).
Thus the proper tags are as such.

Also, when a test is guared by @ver+=X, then the older version
should have exactly(!) @ver-X, and not come up with different version
tags. That is, because we want that there is no overlap and no gap
between the version for which the test is enabled.

[1] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/82009e21d28b5f3d61b9e1e3cebbf2dd94e8f32a
[2] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/4e857e33fca0f6f92fdc378593b7bd102f2a6572

Fixes: f16e556607c8 ('ipv4: adjust version of ipv4_routing_rules_manipulation')