ovs: use "Write dispatcher" step in @ovs_cloned_mac_set_on_iface test
@ovs_cloned_mac_set_on_iface writes a dispatcher script using a shell
command. It relies on the after_scenario hook of @dist to clean that up.

Instead, use the "Write dispatcher" behave step. That one soon will
automatically take care to delete dispatcher files that it creates.
Then @disp no longer will do the cleanup that it does now. Instead,
we use the proper behave step to create and cleanup the file.

While at it, also drop restarting NetworkManager-dispatcher.service.
It's not necessary.