reproducers: don't hide the script filenames in the behave steps
Imagine, you find that a test leaks "10-save-env.sh". So you grep in
NM-ci for that file name. You find it in "contrib/reproducers/repro_1858344.sh"
So you grep for "repro_1858344.sh" to see who calls that. No results.
So you waste some time trying to understand how this works. Eventually,
you find that "1858344" is the term that you can search for, and the
"repro_*.{sh,py}" are a non-obvious naming convention that is required
to follow.

Don't do that. When referring to the reproducer scripts, refer to the
full name. If you don't like the name, then rename the file, but don't
introduce another remapping of the names, don't let the behave step
obfuscate the filename and don't let the behave step add additional
requirements on the pattern of the name.